Rename: class이름 변경, components폴더 생성 및 component파일 이동
